% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Procedure.R
\name{approx_C}
\alias{approx_C}
\title{placeholder}
\usage{
approx_C(matr_dat, fdepths, alpha, B, gamma, grid)
}
\arguments{
\item{matr_dat:}{data in matrix form - each row contains the grid approximations of one observation}

\item{fdepths:}{corresponding depths for the observations}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping}

\item{B:}{number of smoothed bootstrap samples to use}

\item{gamma:}{tuning parameter for smoothed bootstrap}

\item{grid:}{grid used in approximation of matr_dat}
}
\value{
placeholder
}
\description{
placeholder
}
