% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Comparable.R
\name{comparable_obs_finder}
\alias{comparable_obs_finder}
\title{placeholder}
\usage{
comparable_obs_finder(main_interval, measuring_intervals, lambda, ids)
}
\arguments{
\item{main_interval:}{vector of two elements: starting and end point of measuring interval}

\item{measuring_intervals:}{use output from measuring_int()}

\item{lambda:}{acceptable stretching parameter}

\item{:}{ids: identifiers of individual observations}
}
\value{
placeholder
}
\description{
placeholder
}
