% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Stretch.R
\name{dectection_zr_smpl}
\alias{dectection_zr_smpl}
\title{placeholder}
\usage{
dectection_zr_smpl(
  cl,
  list_path,
  measuring_intervals,
  n_obs,
  lambda,
  n_samples = NULL,
  sample_size = NULL,
  alpha = NULL,
  B = NULL,
  gamma = 0.05
)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{measuring_intervals:}{matrix of measuring intervals}

\item{n_obs:}{number of observations in the data set}

\item{lambda:}{acceptable stretching parameter}

\item{n_samples:}{number of samples to use in each iteration 
(NULL for procedure determining value)}

\item{sample_size:}{number of observations to use in each sample in each iteration 
(NULL for procedure determining value)}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping 
(in approximation of C) (NULL for procedure determining value)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C) 
(NULL for procedure determining value)}

\item{gamma:}{tuning parameter for smoothed bootstrap (in approximation of C)}
}
\value{
placeholder
}
\description{
placeholder
}
