% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Wrap.R
\name{detection_wrap}
\alias{detection_wrap}
\title{placeholder}
\usage{
detection_wrap(func_dat, ids, alpha, B, gamma = 0.05)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping 
(in approximation of C - optional if C is specified)}

\item{ids:}{B: number of smoothed bootstrap samples to use 
(in approximation of C - optional if C is specified)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
placeholder
}
\description{
placeholder
}
