% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Grid.R
\name{grid_approx_set_obs}
\alias{grid_approx_set_obs}
\title{placeholder}
\usage{
grid_approx_set_obs(func_dat, grid)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}

\item{grid:}{grid to use for approximation}
}
\value{
placeholder
}
\description{
placeholder
}
