% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Grid.R
\name{grid_finder}
\alias{grid_finder}
\title{placeholder}
\usage{
grid_finder(func_dat)
}
\arguments{
\item{start:}{start point of the grid}

\item{end:}{end point of the grid}
}
\value{
placeholder
}
\description{
placeholder
}
