% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Intervals.R
\name{measuring_int}
\alias{measuring_int}
\title{placeholder}
\usage{
measuring_int(func_dat)
}
\arguments{
\item{func_dat:}{list that contains the observations
e ach observation is a list, that conatins two vectors of identical length: args and vals}
}
\value{
placeholder
}
\description{
placeholder
}
