% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Procedure.R
\name{outlier_detection}
\alias{outlier_detection}
\title{placeholder}
\usage{
outlier_detection(
  matr_dat,
  alpha = 0.05,
  B = 100,
  gamma = 0.05,
  ids,
  grid,
  C = NULL
)
}
\arguments{
\item{matr_dat:}{data in matrix form - each row contains the grid approximations of one observation}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping 
(in approximation of C - optional if C is specified)}

\item{:}{C: should be provided. Otherwise C will be approximated in each step of the iteration}
}
\value{
placeholder
}
\description{
placeholder
}
