% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Sampling.R
\name{sampling_wrap}
\alias{sampling_wrap}
\title{placeholder}
\usage{
sampling_wrap(cl, n_obs, n_samples, sample_size, alpha, B, gamma, list_path)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{n_obs:}{number of observations in data set}

\item{n_samples:}{number of samples to use}

\item{sample_size:}{number of observations to use in each sample}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping 
(in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}

\item{list_path:}{path to the random access list of the data set (generated by package largeList)}
}
\value{
placeholder
}
\description{
placeholder
}
