% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Stretch.R
\name{stretch_and_sample}
\alias{stretch_and_sample}
\title{placeholder}
\usage{
stretch_and_sample(
  cl,
  n_samples,
  sample_size,
  alpha,
  B,
  gamma,
  list_path,
  measuring_interval,
  comparable
)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{n_samples:}{number of samples to use}

\item{sample_size:}{number of observations to use in each sample}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping 
(in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap (in approximation of C)}

\item{list_path:}{path to the random access list of the data set 
(generated by package largeList)}

\item{measuring_interval:}{a vector with 2 elements, 
the start and end points of the desired measuring interval}

\item{comparable:}{vector with the indices of comparable observations in the largelist}
}
\value{
placeholder
}
\description{
placeholder
}
