% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Stretch.R
\name{stretch_data}
\alias{stretch_data}
\title{placeholder}
\usage{
stretch_data(func_dat, measuring_interval)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that conatins two vectors of identical length: args and vals}

\item{measuring_interval:}{a vector with 2 elements, 
the start and end points of the desired measuring interval}
}
\value{
placeholder
}
\description{
placeholder
}
