% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Stretch.R
\name{stretch_obs}
\alias{stretch_obs}
\title{placeholder}
\usage{
stretch_obs(obs, measuring_interval)
}
\arguments{
\item{obs:}{a list that conatins two vectors of identical length: args and vals}

\item{measuring_interval:}{a vector with 2 elements, 
the start and end points of the desired measuring interval}
}
\value{
placeholder
}
\description{
placeholder
}
