% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Intervals.R
\name{unique_intervals}
\alias{unique_intervals}
\title{placeholder}
\usage{
unique_intervals(measuring_intervals)
}
\arguments{
\item{measuring_intervals:}{use output from measuring_int()}
}
\value{
placeholder
}
\description{
placeholder
}
