% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_Shift.R
\name{zero_observations}
\alias{zero_observations}
\title{placeholder}
\usage{
zero_observations(func_dat)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that conatins two vectors of identical length: args and vals}
}
\value{
placeholder
}
\description{
placeholder
}
